Simple implementation of calibration values
The first day / first exercise ask us to extract the first and last
digit from a series of strings, and provide the sum. The example
leaves makes clear that "1", "2", "3", etc are digits but leaves
ambiguous if "one", "two", "three" are also digits. This first
iteration assumes that only numerals are digits and words are not.