Day 3 part 1, part number sum
Interpret the grid of text as a schematic and add up
all the part numbers adjacent to a part symbol.

Adjacency rule includes diagonals.
Part symbol is anything that is not a '.' or a digit.
Account for new line charactrers as well - strip or ignore

This approach uses a regular expression search iterator
followed by an eager search for a part symbol.