Day 3 part 2 find gears
A gear is defined as a part with the '*' symbol and exactly
two part numbers adjacent. The gear ratio is defined as the
product of the part numbers.

This approach scans lines of the schematic similar to part
detection, but this time only finding potential gear parts.
A dictionary with the zero indexed position of the * (stored
as a tuple) and a list of adjacent part numbers is built.

Then this dictionary is iterated over to find proper gears.
Gear ratios are calculated and summed.