Day 4 part 1 score winning scratchers
Given winning numbers and potentially matching game numbers
calculate the score of a series of scratch cards.

This approach puts the winning numbers into a set and the
game numbers into a list. There was some ambiguity regarding
having repitions of game numbers. Then each game number is
compared to the list of winning numbers, advancing the score
of the scratch card when there is a match.

Time complexity is roughly proportional to n which should be
best case - you have to compare each number to determine if it
matches the winning numbers.

If you needed to tightly optimize memory you would use a sorted
array of winning numbers and binary search instead.