Day 4 part 2 recursive winners
Here the scratch cards score not points but recursive copies
of the existing pile of scratch cards. Copies of cards win
more copies of cards. An undefined computational endpoint
is possible depending on input.

I used a stack based recursion approach with a memoization
optimization where the score of each card was calculated
and stored in a dicitonary at the start of the procedure.

This is the first case where I suspect a non-trivial
time complexity optimization to the brute force approach
exists. This took about 10 seconds to compute using this
recursive approach.