Day 5 part 1 decoding sequential maps
The exercise presents us with a series of discrete
maps defined step wise with a series of simple arrithmatic
calculations. The principle here could be applied to more
complex mathematic functions though.

This implementation takes a brute force approach to finding
the map output, by checking each input range until finding
a match. When no match is found the input is the output
as defined in the exercise instructions.

A production version of this code would make a few important
changes. The input ranges are not checked for overlap, and
should be. The linear eager search could be improved to
logarithmic search by storing the imput ranges in a comparable
object and organizing them in a binary search tree. As is
common in situations like this there is no checking for overflow.
Finally, if there is no variation in the desired output - that is
to say if a production system always looked up location (the
final output in the map pairs) then you could compress the map
into a direct lookup instead of a sequental one.