Day 5 part 2 a reverse map brute force approach
This overall approach is to search locations from closest (0)
to furthest, processing the maps in reverse and then checking
if the found seed is one of the issued seeds.

This is still a brute force approach in linear time it works
faster only because the straight forward N is smaller for this
input.

There was a minor improvement of using binary search to find the
appropriate rule in the rules lists instead of linear search, but
this does not improve the problem from linear complexity.

This did compute the correct result somewhere between 1 and 2
hours. I looked up a better solution that was close to one I had
started to consider - breaking the problem down into monotonicly
increasing ranges and then checking only the smallest value
in those. I may or may not implement this improvement later.