Day 6 part 2 winning range search linear approach
A pure brute force approach will clearly take a while here
and is very sub-optimal. This approach finds the range of
winning charges through linear search from each side.

While it completes in a few seconds, this could clearly be improved
with a modified binary search pattern. Time complexity is still
proportional to N but we could get it down to lg N