Day 7 part 1 order winners and score
The exercise is to score poker-like hands of cards and then
order them first by poker categories (2-pair, full house, etc)
then lexicographically where 2-9 are in natural order and
T < J < Q < K < A.

This solution is a bit of a hack that takes advantage of how
Python orders tuples. By storing the hand type or category as
the first element in a tuple representing each hand, and a
translated string as the second element such that the natural
lexicographic ordering is the same as the desired ordering.

A more thorough implementation would use a comparison interface
and a custom class.