Day 7 part 2 Jacks are wild
Minor variation on the existing game, the lexicographic sort
strategy allowed coding wild jacks by simply updating the
translation dictionary and slightly modifying the type
categorization method to count wildcards and usually add that
count to the largest count of pairs, tripples, four-of-a-kind.
Though five-of-a-kind jacks is a degenerate case that was handled
with branch logic.