Day 9 part 1 - sequence prediction
Using a recursive approach, predict the next value in a sequence.
The algorithm for doing this is given as part of the exercise.
Calculate decreasing size difference series until you get to zero.
Then add the end of the last to the end of the previous.
Repeat this in a recursive way.

This implementation uses a stack and is a literal implementation
of the algorithm in the exercise. There's a fairly obvious
memory inefficency - we only care about the last value in each
series but this approach calculates and stores all of them.