day 10 part 1 multiple maze traversals
This exercise asks us to computationally traverse a maze and
find a path that is a loop connected to a given starting point.

This approach creates an object that traverses the maze at each
valid adjacency to the start point. Tiles that are not pipes and
pipes that do not connect to the start are not considered valid
start points for a traversing object.

This satisifies the exercise but has some potential problems:
1. If the input data has two loops connected to the starting
position then only the shortest loop will be detected. The
exercises states that inputs for the exercise are curated to not
contain such a condition but a more complete implementation would
account for potentially exceptional inputs.
2. If the input data has zero loops connected to the starting
position this is not detected. Again the exercise excludes needing
to check for this but a more robust implementation would
3. The readability should be improved by replacing most or all
of the hard coded stuff with constants. Especially the cardinal
step tuples sould be defined as constants such as RIGHT = (1, 0)
because at least two bugs were introduced by not doing this.